fix paths and lint issues after moving bit-javascript content into bit-bin. fixed were done by @GiladShoham in the PR: https://github.com/teambit/bit/pull/2626
