+++ src/consumer/component/dependencies/files-dependency-builder/path-map.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import pathMapFixture from '../../fixtures/path-map.json';
+import pathMapFixture from '../../../../../fixtures/path-map.json';
 import { getPathMapWithLinkFilesData } from './path-map';
 
 describe('path-map', () => {
